Add high LST
