+++ city_metrix/layers/__init__.py
@@ -2,4 +2,5 @@ from .albedo import Albedo
 from .esa_world_cover import EsaWorldCover, EsaWorldCoverClass
 from .land_surface_temperature import LandSurfaceTemperature
 from .tropical_tree_cover import TropicalTreeCover
+from .high_land_surface_temperature import HighLandSurfaceTemperature
 from .layer import Layer, LayerGroupBy
