+++ city_metrix/metrics.py
@@ -1,4 +1,4 @@
-from city_metrix.layers import Albedo, EsaWorldCoverClass, EsaWorldCover, LandSurfaceTemperature, TropicalTreeCover
+from city_metrix.layers import Albedo, EsaWorldCoverClass, EsaWorldCover, HighLandSurfaceTemperature, TropicalTreeCover
 
 
 def built_land_without_tree_cover(zones):
@@ -20,12 +20,17 @@ def surface_reflectivity(zones, start_time, end_time, albedo_threshold):
     built_up_land = EsaWorldCover(land_cover_class=EsaWorldCoverClass.BUILT_UP)
     albedo = Albedo(start_time=start_time, end_time=end_time, threshold=albedo_threshold)
 
-    built_land = built_up_land.groupby(zones).count()
-    built_albedo = albedo.mask(built_land).groupby(zones).count()
+    built_land_counts = built_up_land.groupby(zones).count()
+    built_albedo_counts = albedo.mask(built_up_land).groupby(zones).count()
 
-    return built_albedo / built_land
+    return built_albedo_counts / built_land_counts
 
 
 def high_land_surface_temperature(zones):
-    indicator = LandSurfaceTemperature(hottest_90_days=True).groupby(zones).mean()
-    return indicator
+    built_up_land = EsaWorldCover(land_cover_class=EsaWorldCoverClass.BUILT_UP)
+    high_lst = HighLandSurfaceTemperature()
+
+    built_land_counts = built_up_land.groupby(zones).count()
+    built_high_lst_counts = high_lst.mask(built_up_land).groupby(zones).count()
+
+    return built_high_lst_counts.fillna(0) / built_land_counts
